Add files via upload
Conducted marketing research to study the role of sentiments in user’s tweet sharing behaviour in twitter conditional on industry type and its variation across pre and post covid times. The industries chosen for this study were the Airline and Ecommerce industries. The analysis indicates that effects of sentiment on sharing while not positive in general business as usual scenarios, show variation across industries in times of crisis conditional on customer expectations.